+++ docs/source/changelog.rst
@@ -7,6 +7,7 @@ Unreleased
 .. No unreleased changes.
 
 * Added icon to the "About DMap" menu item in the Help menu.
+* Updated the “Reset View” action: renamed to “Refit View”, updated icon, and now fits the map to the window.
 * Linux release builds are now single-file executables.
 
 :release-link:`v1.0.0`